+++ test/flhooks_test.dart
@@ -2,12 +2,47 @@ import 'package:flhooks/flhooks.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+class _TestHookWidget extends HookWidget {
+  @override
+  Widget builder(BuildContext context) {
+    return Container();
+  }
+}
+
 void main() {
   test('hooks function cannot be called outside HookBuilder', () async {
     expect(() {
       useEffect(() => null, []);
     }, throwsAssertionError);
   });
+  testWidgets('useContext return the actual BuildContext',
+          (WidgetTester tester) async {
+        // You can use keys to locate the widget you need to test
+        var sliderKey = UniqueKey();
+        BuildContext _context = null;
+        BuildContext _contextFromHook = null;
+        // Tells the tester to build a UI based on the widget tree passed to it
+        await tester.pumpWidget(
+          HookBuilder(
+            builder: (BuildContext context) {
+              _context = context;
+              _contextFromHook = useContext();
+              return MaterialApp(
+                home: Material(
+                  child: Center(
+                    child: Slider(
+                      key: sliderKey,
+                      value: 0,
+                      onChanged: (newValue) => null,
+                    ),
+                  ),
+                ),
+              );
+            },
+          ),
+        );
+        expect(_context, _contextFromHook);
+      });
   testWidgets(
       'useState change the state and useCallback memoize the function reference',
       (WidgetTester tester) async {
@@ -127,4 +162,7 @@ void main() {
           ),
         );
       });
+  testWidgets('HookWidget build withour error', (tester) async {
+    tester.pumpWidget(_TestHookWidget());
+  });
 }